# docs/core/specifications/api.openapi.yaml
openapi: 3.1.0
info:
  title: Chit API
  version: 0.2.0
  description: |
    Draft contract. Platform routes live under /api; each module owns /api/<module>/... (see docs/modules/<module>/contracts.md).
    Routes marked x-implemented exist in the hub today (core/server + module routes.py). No authentication is applied (ADR-0007, interim, loopback only).
servers:
  - url: /api
paths:
...
  /kids/phone/settings/{id}:
    put: { operationId: setPhoneSettings, x-implemented: true, x-module: kids/kid-view, description: 'Body {enabled?: boolean, share?: {timetable|stars|reminders|homework|activities|bag|grades|health: boolean}}. Turning it off signs out every phone of that child.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/phone/pairings/{id}:
    delete: { operationId: cancelPhonePairing, x-implemented: true, x-module: kids/kid-view, description: 'Cancel the open pairing of child {id}.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  /kids/phone/devices/{id}:
    delete: { operationId: revokePhone, x-implemented: true, x-module: kids/kid-view, description: 'Remove a paired phone.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/phone/device/chores/{id}/done:
    post: { operationId: deviceTickChore, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Body {done: boolean}. The star stays a parent decision.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/phone/device/homework/{id}/done:
    post: { operationId: deviceTickHomework, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Body {done: boolean}.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  /kids/phone/device/homework/{id}:
    delete: { operationId: deviceRemoveHomework, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Only tasks the child wrote.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...
  /kids/phone/device/bag/items/{id}:
    delete: { operationId: deviceRemoveBagItem, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Only items the child added.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  /kids/phone/device/homework/{id}/dismiss:
    post: { operationId: deviceDismissHomework, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Body {dismissed: boolean}: not relevant, or bring it back.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
  /kids/phone/device/reminders/{id}/state:
    post: { operationId: deviceReminderState, x-implemented: true, x-module: kids/kid-view, description: 'Phone gateway route (/api/kids/phone/device/*, device token). Body {state: done|na|null, date?}: this child''s mark on a reminder for a day from today to a week ahead. Only this phone changes.', parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
...